chore(wallet-core): move PersistentDeviceData types to suite-types
